// tcc-frontend/src/app/model/Result.ts
export class Result {
    data: string = '';
    equipamento: string = '';
    ordem: number = 0;
    pom: string = '';
    prioridade_equipamento: number = 0;
    prioridade_ordem: number = 0;
    prioridade_tam: number = 0;
    tam: string = '';
    valor: number = 0;
    priorizado: string = ''
}